+++ extensions/vscode/src/constants.ts
@@ -1,5 +1,5 @@
 export const templatePackageId = "Plugin.Maui.MVVMExpress.Templates";
-export const templatePackageVersion = "1.0.1";
+export const templatePackageVersion = "1.0.2";
 export const appShortName = "mvvmexpress";
 export const pageShortName = "mvvmexpress-page";
 export const defaultAppName = "MauiApp1";